+++ www/components/Layout/Nav/Nav.module.scss
@@ -5,7 +5,6 @@
   padding: 1em;
   color: var(--lightText);
   background-color: var(--primary);
-  border-bottom: 1px solid var(--border);
   display: flex;
 }
