+++ www/components/pages/Meet/SideBar/SideBarButton/SideBarButton.module.scss
@@ -3,7 +3,6 @@
 .side_bar_button {
   width: 100%;
   padding: .5em 1em;
-  color: colors.$primary;
   text-align: left;
   font-weight: bold;
   margin: 0.25em 0;
