--- www/styles/pages/meeting/[id]/meet.module.css
@@ -1,114 +1,0 @@
-.page_title {
-  width: 100%;
-  margin-left: 1em;
-}
-
-.container {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 100%;
-  min-height: 80vh;
-}
-
-.main_container {
-  width: 60%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border-left: 2px solid var(--border);
-  border-right: 2px solid var(--border);
-  padding: 0em 2em 2em 2em;
-  overflow: scroll;
-}
-
-.main_container::-webkit-scrollbar {
-  width: .6em;
-}
-
-.main_container::-webkit-scrollbar-track {
-  background: var(--background);
-  padding: right 1em;
-}
-
-.main_container::-webkit-scrollbar-button {
-  background: var(--background);
-  color: red;
-}
-
-.main_container::-webkit-scrollbar-corner {
-  background: var(--background);
-}
-
-/* thumb refers to the bar on the slider */
-.main_container::-webkit-scrollbar-thumb {
-  background-image: linear-gradient(var(--secondary), var(--primary));
-  border-radius: 25px;
-}
-
-.main_container h3 {
-  width: 100%;
-}
-
-.side_container {
-  width: 20%;
-  align-items: center;
-  max-width: 25%;
-  display: flex;
-  flex-direction: column;
-  padding-left: 1.5em;
-  padding-right: 1.5em;
-}
-
-.side_container h3 {
-  width: 100%;
-}
-
-.loadingContainer {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.card {
-  width: 100%;
-  padding: 1em;
-  margin: .5em;
-}
-
-.card:hover {
-  cursor: pointer;
-  opacity: 70%;
-}
-
-.card p {
-  width: 100%;
-  margin: 0;
-  white-space: pre-wrap;
-}
-
-.discussion_container {
-  width: 100%;
-}
-
-.discussion_card {
-  padding: 1em;
-  width: 100%;
-}
-
-.discussion_card h3 {
-  width: 100%;
-  margin: 0;
-  text-align: left;
-}
-
-.discussion_card p {
-  width: 100%;
-  text-align: left;
-}
-
-.discussion_card button {
-  margin-left: auto;
-}
